requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
boto3==1.36.18
botocore==1.36.18
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.17.0
fsspec==2025.2.0
ftfy==6.3.1
grpcio==1.67.1
h11==0.14.0
idna==3.10
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
milvus-lite==2.4.11
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
protobuf==5.29.3
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
pymilvus==2.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
s3transfer==0.11.2
scenedetect==0.6.5.2
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
